src/main/java/ru/skypro/exam/kurs_work/service/JavaQuestionService.java
package ru.skypro.exam.kurs_work.service;

import org.springframework.stereotype.Service;
import ru.skypro.exam.kurs_work.exception.QuestionNotFoundException;
import ru.skypro.exam.kurs_work.model.Question;

import java.util.*;

@Service
public class JavaQuestionService implements QuestionService {
    private final Set<Question> questions = new HashSet<>();
    public final Random random = new Random();



    @Override
    public Question add(String question, String answer) {
        return add(new Question(question, answer));
    }

    @Override
    public Question add(Question question) {
        questions.add(question);;
        return question;
    }

    @Override
    public Question remove(Question question) {
        if (!questions.remove(question)) {
            throw new QuestionNotFoundException();
        }
        return question;
    }

    @Override
    public Collection<Question> getAll() {
        return questions;
    }

    @Override
    public Question getRandomQuestion() {
        int questionIndex = random.nextInt(questions.size());
        return new ArrayList<>(questions).get(questionIndex);
    }

}
